refactor: move unparkinorder under sync
